add redhat spec variables for erb templating
This commit swaps out macro assignments and adds templating for use
with the packaging repo. For change visibility, this commit doesn't
is separate from the copy to a .erb, in which git would show the whole
file as a giant change rather than just these changes. Source0 is also
changed to tar.gz, because that is the format that the packaging repo
generates.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>